+++ test/specs/components/PresenceIndicator/PresenceIndicator-test.tsx
@@ -1,7 +1,0 @@
-import { isConformant } from 'test/specs/commonTests'
-
-import PresenceIndicator from 'src/components/PresenceIndicator/PresenceIndicator'
-
-describe('PresenceIndicator', () => {
-  isConformant(PresenceIndicator)
-})